Move card to dedicated directory
