+++ components/Source.js
@@ -2,7 +2,7 @@ import { Dialog } from "@headlessui/react"
 import { useEffect, useState } from "react"
 
 import Button from "@components/Button/Button/Button"
-import Card from "@components/Card"
+import Card from "@components/Card/Card"
 import EmbeddedFrame from "@components/EmbeddedFrame"
 import RequestTooltip from "@components/RequestTooltip"
 import SyncManagement from "@components/SyncManagement"
