+++ pages/integrations/export-crm.js
@@ -2,7 +2,7 @@ import Head from "next/head"
 import { useState } from "react"
 
 import Button from "@components/Button/Button/Button"
-import Card from "@components/Card"
+import Card from "@components/Card/Card"
 import Destination from "@components/Destination"
 import Header from "@components/Structural/Header/Header"
 import { SyncStatus } from "@components/SyncStatus"
